[refactor] Batch 기능 확장으로 인한, 패키지 변경
